+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 `v0.10.0`_ - 00-Unreleased-2022
 -------------------------------
+Deprecated
+++++++++++
+- Support of Python < 3.6 is deprecated and will be removed
+  in next major version.
 
 Removed
 +++++++